Implement the old behavior with new syntax
In other words, when specifying a regex using `:format` we'll convert it
internally to a rule using `:with`.

In other words, the following:

    validates :email, :email_address => {:format => /.+@.+/}

is equivalent to this:

    validates :email, :email_address => {:with => /.+@.+/}

and is processed the same way behind the scenes.

We cannot yet remove `EmailAddress#valid_with_regex?`, unfortunately, as that is
in effect a public method and could be used externally. We should probably
deprecate it eventually.